Fix Coming Eps Page to include shows +/- 1 day for time zone corrections.
More detail...
Episodes airing in a different time zone to the users own can be missing on the Coming Episodes Page. For today, the current day next week, and the last day in the missing episode range; these 3 days (+/- 1 day) can be missing episodes from the display if they air on a different day in their time zone compared to the users current day.

All episodes in the banner layout for today may display as red or green background even if not aired/missing yet because of their airing time. The background colour of the first episode of today in banner view would be shown for all episodes today in this view, since the variable for the episode ending which determines the background colour was only set for the first episode after the today's day header.

Technical detail...
Fix banner view: cur_ep_enddate wasn't set for every episode.
Remove duplicates from SQL result.