Add a link from the footer number of snatched to episode snatched overview page.
The link to the Episode Overview page is available on all pages except on the Episode Overview page.
Change the default state for all check boxes on the Episode Overview page to not checked.
Add validation to Go button to ensure at least one item is checked on Episode Overview page.
Add highlight to current status text in header on Episode Overview page.
Cleanup manageEpisodeStatuses Javascript.