Change overhaul UI of editShow and anime release groups, refactor and simplify code.
Change list order of option on the right of the displayShow page to be mostly inline with the order of options on editShow.
Change legend wording and text colour on the displayShow page.
Add output message if no release group results are available.
Add cleansing of text used in the processes to a add a show.
Add sorting of AniDB available group results.
Add error handling and related UI feedback to reflect result of AniDB communications.
Change description of config/search ignore and require words.