Change to reduce the time taken to "Update shows" with show data.
Change the "Enable IMDb info" option to be under config/General/Interface.
Fix being able to actually turn IMDb off when it has been turned on.
Change to stop displaying IMDb ui features when the option is off.
Change when imdb info is updated...
a) stop updating from IMDb during the scheduled daily updater for every show.
b) update the IMDb info for a show after snatching an episode for it.