Fix Backlog Init
Fix the limited backlog to execute as and when expected and in line with SickBeard.

The backlog scheduled frequency is the higher of either 12 hours or 2 x the "Recent search frequency" + 7 minutes.

On application startup and every backlog scheduled frequency the backlog will either ...
a) do a full run if the number of days passed since the full backlog has reached the setting(1)...
b) do a limited run if the number of days passed since the full backlog is less than setting(1)...
... (1) "Backlog search frequency" in Search Settings.

Change the default full "Backlog search frequency" to 21 days within a range of 2 - 35 (see Search Settings).