Change Nameparser not to query online sources
Nameparser has a parameter try_indexers, which tells Nameparser
to go online while parsing the filename for a show.

Since we have all the information for the shows we want to handle
in our database, there is no need to query indexers for the
information.

This change also reduces abuse of the indexers, since SG no longer
queries indexers when ever an unknown file is postprocessed.