Change if comparison from = to == for POSIX compliance
While `==` looks nicer (more obvious at a glance that it's a comparison, rather than an assignment), unfortunately it's a bit of a Bashism. Many modern shells support it, but not all do, particularly deliberately minimal shells---which includes for instance Dash, which daemon users on Debian and derivatives like Ubuntu use by default (which is how I ran face-first into this myself). The base POSIX comparison operand is merely `=`, and that should work everywhere, then.

References: 
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/test.html#tag_20_128_05
https://mywiki.wooledge.org/Bashism#Conditionals