Change couple more consistency improvements.
Change undo one simplification too far. Reasoning, the var is used later on in code, the natural eyeline is to look for the variable defined on the left above, and the declaration was hidden mid line due to walrus. Therefore, this makes it take longer to scan over code and slows down readability, i Bet this will be written about in some blog very soon.
Add ranking to scores on card
Change use new refactored overview cleaner